requirements.txt
requests==2.31.0
html2text==2024.2.26
python-dotenv==1.0.0
